+++ test/src/features/return_result/repositories/return_result_repository_test.dart
@@ -19,6 +19,7 @@ void main() {
     ),
     nameFeature: "Teste Usecase",
     showRuntimeMilliseconds: true,
+    isIsolate: false,
   );
 
   setUp(() {