+++ test/src/interfaces/usecase_implement_test.dart
@@ -33,6 +33,7 @@ void main() {
     ),
     nameFeature: "Teste Usecase",
     showRuntimeMilliseconds: true,
+    isIsolate: false,
   );
 
   setUp(() {